rename toplevel checker module to "checker", add CLI option.
