+++ src/vm/checker/errors.rs
