+++ src/vm/checker/mod.rs
